+++ internal/compiler/llr/expression.rs
@@ -195,6 +195,7 @@ impl Expression {
             | Type::Angle
             | Type::PhysicalLength
             | Type::LogicalLength
+            | Type::Rem
             | Type::UnitProduct(_) => Expression::NumberLiteral(0.),
             Type::Percent => Expression::NumberLiteral(1.),
             Type::String => Expression::StringLiteral(String::new()),